Code review comments for spoke cluster oauth clients
